pyproject.toml
[project]
name = "laudareproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "keras>=3.12.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.90",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.1",
]
